Use LL/macros for clock setup
Also fixed the formatting and removed the old MX file (and a
peripheral definitions file which I made, and I think it was
for things like "huart", but they actually don't exist for
LL functions, you just specify which peripheral you're
using. It's great.)

Rather than remove the lines to compile the full HAL from
the library makefile, since it's not just "glob all" because
templates mess things up, I made it switchable.